+++ frontend/src/components/features/sidebar/sidebar.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
-import { LoadingSpinner } from "#/components/modals/loading-project";
 import { useAuth } from "#/context/auth-context";
 import { useUserPrefs } from "#/context/user-prefs-context";
 import { useGitHubUser } from "#/hooks/query/use-github-user";
@@ -8,11 +7,12 @@ import { useIsAuthed } from "#/hooks/query/use-is-authed";
 import { AllHandsLogoButton } from "#/components/ui/buttons/all-hands-logo-button";
 import { DocsButton } from "#/components/ui/buttons/docs-button";
 import { SettingsButton } from "#/components/ui/buttons/settings-button";
-import { AccountSettingsModal } from "#/components/account-settings-modal";
-import { ExitProjectConfirmationModal } from "#/components/exit-project-confirmation-modal";
-import { SettingsModal } from "#/components/settings-modal";
+import { AccountSettingsModal } from "#/components/modals/account-settings/account-settings-modal";
+import { SettingsModal } from "#/components/modals/settings/settings-modal";
 import { UserActions } from "./user-actions";
 import { ExitProjectButton } from "../../ui/buttons/exit-project-button";
+import { ExitProjectConfirmationModal } from "#/components/modals/exit-project-confirmation-modal";
+import { LoadingSpinner } from "#/components/ui/loading-spinner";
 
 export function Sidebar() {
   const location = useLocation();
