+++ frontend/src/components/features/sidebar/user-avatar.tsx
@@ -1,7 +1,7 @@
 import DefaultUserAvatar from "#/icons/default-user.svg?react";
 import { cn } from "#/utils/utils";
 import { Avatar } from "./avatar";
-import { LoadingSpinner } from "../../modals/loading-project";
+import { LoadingSpinner } from "../../ui/loading-spinner";
 
 interface UserAvatarProps {
   onClick: () => void;
