+++ frontend/src/components/modals/account-settings/account-settings-modal.tsx
@@ -1,4 +1,4 @@
-import { AccountSettingsForm } from "#/components/modals/account-settings-form";
+import { AccountSettingsForm } from "#/components/modals/account-settings/account-settings-form";
 import { ModalBackdrop } from "#/components/modals/modal-backdrop";
 import { useUserPrefs } from "#/context/user-prefs-context";
 import { useGitHubUser } from "#/hooks/query/use-github-user";
