+++ frontend/src/components/modals/confirmation-modals/danger-modal.tsx
@@ -1,4 +1,4 @@
-import BaseModal from "./base-modal";
+import { BaseModal } from "./base-modal";
 
 interface DangerModalProps {
   title: string;
