+++ frontend/src/components/modals/exit-project-confirmation-modal.tsx
