+++ frontend/src/components/modals/security/security.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 import SecurityInvariant from "./invariant/invariant";
-import BaseModal from "../base-modal/base-modal";
 import { I18nKey } from "#/i18n/declaration";
+import { BaseModal } from "../base-modal/base-modal";
 
 interface SecurityProps {
   isOpen: boolean;
