+++ frontend/src/components/ui/loading-spinner.tsx
@@ -1,7 +1,23 @@
-export function LoadingSpinner() {
+import LoadingSpinnerOuter from "#/icons/loading-outer.svg?react";
+import { cn } from "#/utils/utils";
+
+interface LoadingSpinnerProps {
+  size: "small" | "large";
+}
+
+export function LoadingSpinner({ size }: LoadingSpinnerProps) {
+  const sizeStyle =
+    size === "small" ? "w-[25px] h-[25px]" : "w-[50px] h-[50px]";
+
   return (
-    <div className="flex justify-center">
-      <div className="w-6 h-6 border-2 border-t-[4px] border-primary-500 rounded-full animate-spin" />
+    <div data-testid="loading-spinner" className={cn("relative", sizeStyle)}>
+      <div
+        className={cn(
+          "rounded-full border-4 border-[#525252] absolute",
+          sizeStyle,
+        )}
+      />
+      <LoadingSpinnerOuter className={cn("absolute animate-spin", sizeStyle)} />
     </div>
   );
 }
